+++ stacks/list.go
@@ -10,7 +10,7 @@ import (
 	"github.com/Scalingo/cli/config"
 )
 
-func List(ctx context.Context) error {
+func List(ctx context.Context, isWithDeprecatedFlag bool) error {
 	c, err := config.ScalingoClient(ctx)
 	if err != nil {
 		return errgo.Notef(err, "fail to get Scalingo client")
@@ -22,14 +22,35 @@ func List(ctx context.Context) error {
 	}
 
 	t := tablewriter.NewWriter(os.Stdout)
-	t.SetHeader([]string{"ID", "Name", "Description", "Default?"})
+
+	if isWithDeprecatedFlag {
+		t.SetHeader([]string{"ID", "Name", "Description", "Default?", "Deprecated?", "Deprecation date"})
+	} else {
+		t.SetHeader([]string{"ID", "Name", "Description", "Default?"})
+	}
 
 	for _, stack := range stacks {
 		defaultText := "No"
 		if stack.Default {
 			defaultText = "Yes"
 		}
-		t.Append([]string{stack.ID, stack.Name, stack.Description, defaultText})
+
+		if isWithDeprecatedFlag {
+			deprecatedText := "No"
+			deprecationDate := ""
+
+			if stack.IsDeprecated() {
+				deprecatedText = "Yes"
+			}
+
+			if !stack.DeprecatedAt.IsZero() {
+				deprecationDate = stack.DeprecatedAt.Format("2006-01-02")
+			}
+
+			t.Append([]string{stack.ID, stack.Name, stack.Description, defaultText, deprecatedText, deprecationDate})
+		} else if !stack.IsDeprecated() {
+			t.Append([]string{stack.ID, stack.Name, stack.Description, defaultText})
+		}
 	}
 	t.Render()
 	return nil